Extract ResourceNotFoundFailure into resource_not_found_failure.dart
