+++ lib/src/sdk/errors.dart
@@ -1,12 +1,5 @@
 import 'failure.dart';
 
-class ResourceNotFoundFailure extends Failure {
-  ResourceNotFoundFailure({
-    required String resource,
-    Map<String, String>? context,
-  }) : super(message: 'Resource not found: $resource', context: context);
-}
-
 class UsageFailure extends Failure {
   UsageFailure({
     required String usage,
