+++ lib/src/sdk/resource_not_found_failure.dart
@@ -1,0 +1,8 @@
+import 'failure.dart';
+
+class ResourceNotFoundFailure extends Failure {
+  ResourceNotFoundFailure({
+    required String resource,
+    Map<String, String>? context,
+  }) : super(message: 'Resource not found: $resource', context: context);
+}
